Fnc: Medicap : permettre plusieurs établissements SHS sur un seul établissement Mediboard Fnc: Medicap : permettre de filtrer sur un établissement dans les paramètres GET pour le traitement des triggers
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@6963 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPplanningOp/sejour.class.php
@@ -437,9 +437,13 @@ class CSejour extends CCodable {
   
   function delAffectations() {
     $this->loadRefsAffectations();
-    foreach($this->_ref_affectations as $key => $value) {
-      $this->_ref_affectations[$key]->deleteOne();
-    }
+    
+		// dPhospi might not be active
+		if ($this->_ref_affectations) {
+	    foreach($this->_ref_affectations as $key => $value) {
+	      $this->_ref_affectations[$key]->deleteOne();
+	    }
+		}
     return null;
   }
